+++ frontend/components/component/markerEditor.tsx
@@ -1,31 +1,30 @@
-/**
- * This code was generated by v0 by Vercel.
- * @see https://v0.dev/t/brWdjky99ni
- */
-import { Input } from "@/components/ui/input"
-import { Button } from "@/components/ui/button"
+import { Input } from "@/components/ui/input";
+import { Button } from "@/components/ui/button";
 import { CloseIcon } from "../ui/icons";
-import { Textarea } from "@/components/ui/textarea"
+import { Textarea } from "@/components/ui/textarea";
 import React, { useState } from "react";
+import { MapMarker } from "../types/MapMarker";
 
-export function EditMarker({ onClose, onTitleChange, title }: {
-  onClose: () => void;
-  onTitleChange: (newTitle: string) => void;
-  title: string;
-}) {
+type MarkerEditorType = {
+  onClose: () => void,
+  onTitleChange: (newTitle: string) => void,
+  marker: MapMarker | null,
+}
 
-  const [inputTitle, setInputTitle] = useState(title);
+export const MarkerEditor = (props: MarkerEditorType) => {
+
+  const [inputTitle, setInputTitle] = useState(props.marker?.display_name ?? "");
 
   const handleUpdate = () => {
-    onTitleChange(inputTitle);
-    onClose();
+    props.onTitleChange(inputTitle);
+    props.onClose();
   }
 
   return (
     <div className="fixed inset-0 bg-gray-500 bg-opacity-75 flex items-center justify-center p-4">
       <div className="bg-white dark:bg-gray-800 rounded-lg max-w-2xl w-full p-6 space-y-4 relative">
         <button className="absolute top-4 right-4">
-          <CloseIcon className="h-6 w-6" onClick={onClose} />
+          <CloseIcon className="h-6 w-6" onClick={props.onClose} />
         </button>
         <h2 className="text-xl font-semibold">Edit location</h2>
         <div className="flex flex-col space-y-3">
@@ -38,7 +37,7 @@ export function EditMarker({ onClose, onTitleChange, title }: {
           <Button
             className="w-1/2"
             variant={"secondary"}
-            onClick={onClose}>
+            onClick={props.onClose}>
             Cancel
           </Button>
           <Button
